Finished a first pass at the correlator log trade extraction script
The extraction script now properly identifies full and partial fills from the extraParams part of the OrderUpdate entry and pulls out the fills. The output agrees with the output from the FIX logs.

Unfortunately, this does assume that the order management is communicating via FIX, so if we use other protocols, it may be necessary to make alterations to the extraction logic.

git-svn-id: file:///var/lib/svn/repository/sysadmin/trunk@83 f22ea994-e202-4195-b786-b54d0d3ddba0